implement suggestions

+++ src/core/control/xojfile/SaveHandler.cpp
@@ -267,7 +267,7 @@ void SaveHandler::visitPage(XmlNode* root, ConstPageRef p, const Document* doc, 
             }
         }
         background->setAttrib("pageno", p->getPdfPageNr() + 1);
-        background->setAttrib("pageorient", p->getPdfPageOrientation());
+        background->setAttrib("pageorient", static_cast<int>(p->getPdfPageOrientation()));
     } else if (p->getBackgroundType().isImagePage()) {
         background->setAttrib("type", "pixmap");
